refactor: Move routes to /routes
