+++ app/templates/index.html
@@ -26,8 +26,8 @@
             </div>
             
             <div class="nav-right">
-                <a href="/auto-translation" class="nav-button">Translate</a>
-                <a href="/auto-review" class="nav-button">Review</a>
+                <a href="/translate" class="nav-button">Translate</a>
+                <a href="/review" class="nav-button">Review</a>
                 <a href="/commit" class="nav-button">Commit</a>
                 <a href="/publish" class="nav-button">Publish</a>
